requirements.txt
torch==1.13.0
torchvision==0.14.0
pytorch_lightning==1.9.0
Pillow==10.0.0
opencv-python
transformers
taming-transformers-rom1504 
tqdm
numpy==1.23
kornia
webdataset
omegaconf
einops
scikit-image
pymcubes
carvekit-colab
open3d
trimesh
easydict
imageio-ffmpeg==0.4.7
git+https://github.com/openai/CLIP.git
sk-video
moviepy
